deleted Player as entety and added it as class
