+++ src/main/java/ch/uzh/ifi/hase/soprafs23/custom/Player.java
@@ -1,0 +1,36 @@
+package ch.uzh.ifi.hase.soprafs23.custom;
+
+import javax.persistence.Embeddable;
+
+@Embeddable
+public class Player {
+    private String name;
+    private int personalScore;
+    private boolean leader;
+
+
+
+    public int getPersonalScore() {
+        return personalScore;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setLeader(boolean leader) {
+        this.leader = leader;
+    }
+
+    public boolean isLeader() {
+        return leader;
+    }
+
+    public void setPersonalScore(int personalScore) {
+        this.personalScore = personalScore;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
