--- src/main/java/ch/uzh/ifi/hase/soprafs23/entity/Player.java
@@ -1,67 +1,0 @@
-package ch.uzh.ifi.hase.soprafs23.entity;
-
-import ch.uzh.ifi.hase.soprafs23.constant.UserStatus;
-
-import javax.persistence.*;
-import java.io.Serializable;
-
-/**
- * Internal User Representation
- * This class composes the internal representation of the user and defines how
- * the user is stored in the database.
- * Every variable will be mapped into a database field with the @Column
- * annotation
- * - nullable = false -> this cannot be left empty
- * - unique = true -> this value must be unqiue across the database -> composes
- * the primary key
- */
-@Entity
-@Table(name = "PLAYER")
-public class Player implements Serializable {
-
-
-  @Id
-  @GeneratedValue
-  private Long id;
-
-  @Column(nullable = false, unique = true)
-  private String username;
-
-  @Column(nullable = false)
-  private int personalScore;
-
-  @Column(nullable = false)
-  private boolean isLeader;
-
-  public Long getId() {
-    return id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getUsername() {
-    return username;
-  }
-
-  public void setUsername(String username) {
-    this.username = username;
-  }
-
-    public int getPersonalScore() {
-        return personalScore;
-    }
-
-    public void setPersonalScore(int personalScore) {
-        this.personalScore = personalScore;
-    }
-
-    public void setLeader(boolean isLeader) {
-        this.isLeader = isLeader;
-    }
-
-    public boolean isLeader() {
-        return isLeader;
-    }
-}
