+++ src/main/java/ch/uzh/ifi/hase/soprafs23/rest/dto/TeamGetDTO.java
@@ -1,7 +1,7 @@
 package ch.uzh.ifi.hase.soprafs23.rest.dto;
 
 import ch.uzh.ifi.hase.soprafs23.constant.Role;
-import ch.uzh.ifi.hase.soprafs23.entity.Player;
+import ch.uzh.ifi.hase.soprafs23.custom.Player;
 
 import java.util.List;
